requirements.txt
imgaug==0.4.0
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.6.0.66
Pillow==9.2.0
PyYAML==6.0
scikit_learn==1.1.1
scipy==1.8.0
torch
torchvision
tqdm==4.64.0
